requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Date/time handling
pytz>=2023.3

# GPX file handling (for activity exports)
gpxpy>=1.5.0

# Progress bars
tqdm>=4.65.0

# Configuration
click>=8.1.0

# Testing (dev dependency)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code formatting (dev dependency)
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
